No need for unsigned long long

+++ src/CalcViewModel/HistoryViewModel.h
@@ -53,7 +53,7 @@ namespace CalculatorApp
 
             internal : HistoryViewModel(_In_ CalculationManager::CalculatorManager* calculatorManager);
             void SetCalculatorDisplay(Common::CalculatorDisplay& calculatorDisplay);
-            unsigned long long GetMaxItemSize();
+            size_t GetMaxItemSize();
 
 
         private:
